requirements.txt
aiohttp==3.9.2
async-timeout==4.0.2
attrs==23.1.0
cffi==1.15.1
charset-normalizer==3.2.0
discord==2.3.1
discord.py==2.3.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
tzdata==2023.3
yarl==1.9.2
youtube-dl==2021.12.17
